Split `JS` singleton class out of `tools/shared.py`. NFC
Just a refactoring to reduce the size of shared.py and the
pointless use of a class/nesting.
