+++ emcc.py
@@ -954,6 +954,20 @@ def move_file(src, dst):
   shutil.move(src, dst)
 
 
+# Returns the subresource location for run-time access
+def get_subresource_location(path, data_uri=None):
+  if data_uri is None:
+    data_uri = settings.SINGLE_FILE
+  if data_uri:
+    # if the path does not exist, then there is no data to encode
+    if not os.path.exists(path):
+      return ''
+    data = base64.b64encode(utils.read_binary(path))
+    return 'data:application/octet-stream;base64,' + data.decode('ascii')
+  else:
+    return os.path.basename(path)
+
+
 run_via_emxx = False
 
 
@@ -2337,7 +2351,7 @@ def phase_linker_setup(options, state, newargs, settings_map):
     if settings.SINGLE_FILE:
       exit_with_error('NODE_CODE_CACHING saves a file on the side and is not compatible with SINGLE_FILE')
 
-  if not shared.JS.isidentifier(settings.EXPORT_NAME):
+  if not js_manipulation.isidentifier(settings.EXPORT_NAME):
     exit_with_error(f'EXPORT_NAME is not a valid JS identifier: `{settings.EXPORT_NAME}`')
 
   if settings.EMSCRIPTEN_TRACING and settings.ALLOW_MEMORY_GROWTH:
@@ -2728,7 +2742,7 @@ def phase_final_emitting(options, state, target, wasm_target, memfile):
   if settings.MINIMAL_RUNTIME == 2 and settings.USE_CLOSURE_COMPILER and settings.DEBUG_LEVEL == 0 and not settings.SINGLE_FILE:
     # Process .js runtime file. Note that we need to handle the license text
     # here, so that it will not confuse the hacky script.
-    shared.JS.handle_license(final_js)
+    js_manipulation.handle_license(final_js)
     shared.run_process([shared.PYTHON, utils.path_from_root('tools/hacky_postprocess_around_closure_limitations.py'), final_js])
 
   # Unmangle previously mangled `import.meta` references in both main code and libraries.
@@ -2750,7 +2764,7 @@ def phase_final_emitting(options, state, target, wasm_target, memfile):
       f.write(options.extern_post_js)
     save_intermediate('extern-pre-post')
 
-  shared.JS.handle_license(final_js)
+  js_manipulation.handle_license(final_js)
 
   js_target = state.js_target
 
@@ -3308,7 +3322,7 @@ def phase_binaryen(target, options, wasm_target):
     if settings.MINIMAL_RUNTIME:
       js = do_replace(js, '<<< WASM_BINARY_DATA >>>', base64_encode(read_binary(wasm_target)))
     else:
-      js = do_replace(js, '<<< WASM_BINARY_FILE >>>', shared.JS.get_subresource_location(wasm_target))
+      js = do_replace(js, '<<< WASM_BINARY_FILE >>>', get_subresource_location(wasm_target))
     shared.try_delete(wasm_target)
     write_file(final_js, js)
 
@@ -3424,7 +3438,7 @@ def generate_traditional_runtime_html(target, options, js_target, target_basenam
   var filename = '%s';
   if ((',' + window.location.search.substr(1) + ',').indexOf(',noProxy,') < 0) {
     console.log('running code in a web worker');
-''' % shared.JS.get_subresource_location(proxy_worker_filename)) + worker_js + '''
+''' % get_subresource_location(proxy_worker_filename)) + worker_js + '''
   } else {
     console.log('running code on the main thread');
     var fileBytes = tryParseAsDataURI(filename);
@@ -3453,7 +3467,7 @@ def generate_traditional_runtime_html(target, options, js_target, target_basenam
           meminitXHR.open('GET', memoryInitializer, true);
           meminitXHR.responseType = 'arraybuffer';
           meminitXHR.send(null);
-''' % shared.JS.get_subresource_location(memfile)) + script.inline
+''' % get_subresource_location(memfile)) + script.inline
 
     if not settings.WASM_ASYNC_COMPILATION:
       # We need to load the wasm file before anything else, it has to be synchronously ready TODO: optimize
@@ -3475,7 +3489,7 @@ def generate_traditional_runtime_html(target, options, js_target, target_basenam
 %s
           };
           wasmXHR.send(null);
-''' % (shared.JS.get_subresource_location(wasm_target), script.inline)
+''' % (get_subresource_location(wasm_target), script.inline)
 
     if settings.WASM == 2:
       # If target browser does not support WebAssembly, we need to load the .wasm.js file before the main .js file.
@@ -3495,7 +3509,7 @@ def generate_traditional_runtime_html(target, options, js_target, target_basenam
             // Current browser supports Wasm, proceed with loading the main JS runtime.
             loadMainJs();
           }
-''' % (script.inline, shared.JS.get_subresource_location(wasm_target) + '.js')
+''' % (script.inline, get_subresource_location(wasm_target) + '.js')
 
   # when script.inline isn't empty, add required helper functions such as tryParseAsDataURI
   if script.inline:
@@ -3592,7 +3606,7 @@ def generate_html(target, options, js_target, target_basename,
 def generate_worker_js(target, js_target, target_basename):
   # compiler output is embedded as base64
   if settings.SINGLE_FILE:
-    proxy_worker_filename = shared.JS.get_subresource_location(js_target)
+    proxy_worker_filename = get_subresource_location(js_target)
 
   # compiler output goes in .worker.js file
   else:
